updated ddls to be nullable, added EduProfiles and Histories, changed age to DoB on Profile entity, added getCounties stdOption method, decoupled eduHistory from school entities, added relavent options to eduProfile, optimized db relationships and added a school address entity, mostly completed eduProfile, began work on EduHistory, added collapsable bootstap containers to seperate visually partials
really really tried to attempt to use 1:1 relationships in EF but doing so, over several different instances, dramatically increased code complexity. Instead I opted to keep 1:0..1 (I think that is the relationship in use) as it meshes better with how I have built conventions. The biggest issue with teh 1:1 was the base class.